# tests/snapshots/openapi-v1.json
{
 "components": {
  "schemas": {
...
   "ForecastIn": {
    "properties": {
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "org": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Org"
     },
     "privacy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Privacy"
     },
     "project": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project"
     },
     "yaml": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Yaml"
     }
    },
    "required": [
     "goal"
    ],
    "title": "ForecastIn",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "MemoryIn": {
    "properties": {
     "pinned": {
      "default": false,
      "title": "Pinned",
      "type": "boolean"
     },
     "private": {
      "default": false,
      "title": "Private",
      "type": "boolean"
     },
     "scope": {
      "default": "global",
      "title": "Scope",
      "type": "string"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "title": "Tags",
      "type": "array"
     },
     "text": {
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "MemoryIn",
    "type": "object"
   },
...
   "RunIn": {
    "properties": {
     "allow_dirty": {
      "default": false,
      "title": "Allow Dirty",
      "type": "boolean"
     },
     "allow_exec": {
      "anyOf": [
       {
        "const": "container_only",
        "type": "string"
       },
       {
        "type": "boolean"
       }
      ],
      "default": false,
      "title": "Allow Exec"
     },
     "auto_approve": {
      "default": false,
      "title": "Auto Approve",
      "type": "boolean"
     },
     "base": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base"
     },
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "org": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Org"
     },
     "privacy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Privacy"
     },
     "project": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project"
     },
     "team": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team"
     },
     "yaml": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Yaml"
     }
    },
    "required": [
     "goal"
    ],
    "title": "RunIn",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "Cadre"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/approvals": {
   "get": {
    "operationId": "approvals_api_v1_approvals_get",
    "parameters": [
     {
      "in": "query",
      "name": "pending",
      "required": false,
      "schema": {
       "default": true,
       "title": "Pending",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "run",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Approvals Api V1 Approvals Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Approvals"
   }
  },
...
  "/api/v1/forecast": {
   "post": {
    "operationId": "forecast_endpoint_api_v1_forecast_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ForecastIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Forecast Endpoint Api V1 Forecast Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Forecast Endpoint"
   }
  },
...
  "/api/v1/memory": {
   "get": {
    "operationId": "memory_list_api_v1_memory_get",
    "parameters": [
     {
      "in": "query",
      "name": "scope",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Scope"
      }
     },
     {
      "in": "query",
      "name": "pending",
      "required": false,
      "schema": {
       "default": true,
       "title": "Pending",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Memory List Api V1 Memory Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Memory List"
   },
   "post": {
    "operationId": "memory_add_api_v1_memory_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MemoryIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Memory Add Api V1 Memory Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Memory Add"
   }
  },
...
  "/api/v1/runs": {
   "get": {
    "operationId": "runs_api_v1_runs_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Runs Api V1 Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Runs"
   },
   "post": {
    "operationId": "start_run_api_v1_runs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Start Run Api V1 Runs Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start Run"
   }
  },
  "/api/v1/runs/{rid}": {
   "get": {
    "operationId": "run_api_v1_runs__rid__get",
    "parameters": [
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Run Api V1 Runs  Rid  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run"
   }
  },
...
  "/api/v1/runs/{rid}/events": {
   "get": {
    "description": "`tail=N` returns the last N events instead of paging forward from `after`.",
    "operationId": "events_api_v1_runs__rid__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "after",
      "required": false,
      "schema": {
       "default": 0,
       "title": "After",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 500,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "tail",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Tail",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Events Api V1 Runs  Rid  Events Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Events"
   }
  },
...
  "/api/v1/runs/{rid}/org": {
   "get": {
    "operationId": "run_org_api_v1_runs__rid__org_get",
    "parameters": [
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run Org"
   }
  },
...
  "/api/v1/usage": {
   "get": {
    "operationId": "usage_api_v1_usage_get",
    "parameters": [
     {
      "in": "query",
      "name": "days",
      "required": false,
      "schema": {
       "default": 7,
       "title": "Days",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Usage Api V1 Usage Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Usage"
   }
  },
...
 }
}